Added splash screen
Edit mainactivity to homeactivity
now mainactivity is the splashscreen
