+++ .idea/misc.xml
@@ -7,7 +7,7 @@
         <entry key="..\:/Users/Luca24/AndroidStudioProjects/PDM_Project/app/src/main/res/layout/activity_about.xml" value="0.33" />
         <entry key="..\:/Users/Luca24/AndroidStudioProjects/PDM_Project/app/src/main/res/layout/activity_info.xml" value="0.24909420289855072" />
         <entry key="..\:/Users/Luca24/AndroidStudioProjects/PDM_Project/app/src/main/res/layout/activity_main.xml" value="0.25" />
-        <entry key="..\:/Users/Luca24/AndroidStudioProjects/PDM_Project/app/src/main/res/layout/activity_splash_screen.xml" value="0.17426545086119555" />
+        <entry key="..\:/Users/Luca24/AndroidStudioProjects/PDM_Project/app/src/main/res/layout/activity_splash_screen.xml" value="0.25" />
         <entry key="..\:/Users/Luca24/AndroidStudioProjects/PDM_Project/app/src/main/res/layout/fragment_first.xml" value="0.20833333333333334" />
         <entry key="..\:/Users/Luca24/AndroidStudioProjects/PDM_Project/app/src/main/res/layout/fragment_info.xml" value="0.24909420289855072" />
         <entry key="..\:/Users/Luca24/AndroidStudioProjects/PDM_Project/app/src/main/res/layout/fragment_second.xml" value="0.20833333333333334" />
