+++ app/src/main/java/com/example/pdm_project/HomeActivity.java
@@ -1,0 +1,77 @@
+package com.example.pdm_project;
+
+import androidx.annotation.NonNull;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatDelegate;
+import androidx.navigation.NavController;
+import androidx.navigation.Navigation;
+import androidx.navigation.ui.AppBarConfiguration;
+import androidx.navigation.ui.NavigationUI;
+
+import android.content.Intent;
+import android.os.Bundle;
+import android.view.Menu;
+import android.view.MenuInflater;
+import android.view.MenuItem;
+import android.widget.CalendarView;
+import android.widget.Toast;
+
+import com.google.android.material.bottomnavigation.BottomNavigationView;
+
+public class HomeActivity extends AppCompatActivity {
+
+    public CalendarView calendario;
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_main);
+
+        BottomNavigationView bottomNavigationView = findViewById(R.id.bottomNavigationView);
+        NavController navController = Navigation.findNavController(this,  R.id.FragmentContainerView);
+        NavigationUI.setupWithNavController(bottomNavigationView, navController);
+
+        AppBarConfiguration appBarConfiguration = new AppBarConfiguration.Builder(R.id.firstFragment, R.id.secondFragment, R.id.thirdFragment).build();
+        NavigationUI.setupActionBarWithNavController(this, navController, appBarConfiguration);
+
+        calendario = findViewById(R.id.calendarView);
+        loadCalendar(calendario);
+   }
+    public void loadCalendar(CalendarView calendarView){
+
+    }
+    @Override
+    public boolean onCreateOptionsMenu(Menu menu) {
+        MenuInflater inflater = getMenuInflater();
+        inflater.inflate(R.menu.right_menu, menu);
+        return true;
+    }
+
+    @Override
+    public boolean onOptionsItemSelected(@NonNull MenuItem item) {
+        switch (item.getItemId()){
+
+            case R.id.menuRightMode:
+                if (getDelegate().getLocalNightMode() == AppCompatDelegate.MODE_NIGHT_YES){
+                    getDelegate().setLocalNightMode(AppCompatDelegate.MODE_NIGHT_NO);
+                    showMessage("Tema Light impostato!");
+                } else{
+                    getDelegate().setLocalNightMode(AppCompatDelegate.MODE_NIGHT_YES);
+                    showMessage("Tema Night impostato!");
+                }
+                return true;
+            case R.id.menuRightInfo:
+                startActivity(new Intent(HomeActivity.this, AboutActivity.class));
+                //showMessage("INFO");
+                return true;
+            default:
+                return super.onOptionsItemSelected(item);
+        }
+
+    }
+
+    private void showMessage(String message){
+        Toast.makeText(this, message, Toast.LENGTH_SHORT).show();
+    }
+
+}
