--- app/src/main/java/com/example/pdm_project/SplashScreen.java
@@ -1,14 +1,0 @@
-package com.example.pdm_project;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.os.Bundle;
-
-public class SplashScreen extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_splash_screen);
-    }
-}
